refactor: 필요없는 로직 제거 및 패키지 이동
