+++ src/main/java/chess/domain/game/User.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 public class User {
