+++ src/test/java/chess/domain/ChessGameTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import chess.domain.game.ChessGame;
 import chess.domain.square.File;
 import chess.domain.square.Rank;
 import chess.domain.square.Square;
